% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na.R
\name{all_na}
\alias{all_na}
\title{A function to return TRUE for columns of a data frame
 which are all NA.}
\usage{
all_na(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A logical vector of length ncol(df), that is TRUE only for columns
where all rows are NA.
}
\description{
This can be useful for removing missing columns, which may occur while cleaning
a data set
}
\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
