% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alp_theme.R
\name{alp_theme}
\alias{alp_theme}
\title{ALP custom theme for ggplot figures}
\usage{
alp_theme()
}
\value{
None
}
\description{
Based on theme_linedraw, with adjustments to axis text, strip format,
and no legends. Any of these formats can be overriden by adding them
to a plot manually. Function works best with theme_set.
}
\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
\keyword{figure,}
\keyword{ggplot,}
\keyword{plot,}
\keyword{theme}
