% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdog.R
\name{catdog}
\alias{catdog}
\title{A wrapper to return head() and tail() of an R object in a list.
 Will inherit S3 methods from the parent functions.}
\usage{
catdog(df, n = 6L)
}
\arguments{
\item{df}{A data frame.}

\item{n}{The number of rows to show from the start and end of the data frame}
}
\value{
A named list of length 2 with elements HEAD and TAIL, consisting of
 the head and tail of the data frame, of length n.
}
\description{
A wrapper to return head() and tail() of an R object in a list.
 Will inherit S3 methods from the parent functions.
}
\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
