% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_na.R
\name{mean_na}
\alias{mean_na}
\alias{sd_na}
\alias{med_na}
\alias{IQR_na}
\alias{sum_na}
\title{Wrappers on base and stats functions where na.rm = TRUE automatically.}
\usage{
mean_na(x)

sd_na(x)

med_na(x)

IQR_na(x)

sum_na(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A numeric value resulting from the chosen statistical operation,
ignoring NAs.
}
\description{
Useful for data sets where you don't want to get rid of the NAs but need
to calculate summary stats without them. Especially useful when you get
tired of typing na.rm = TRUE over and over and over...
}
\section{mean_na}{

Calculates the mean of a numeric vector input, with na.rm = TRUE.
}

\section{sd_na}{

Calculates the standard deviation of a numeric vector input, with na.rm = TRUE.
}

\section{med_na}{

Calculates the median of a numeric vector input, with na.rm = TRUE.
}

\section{IQR_na}{

Calculates the inter quartile range of a numeric vector input, with na.rm = TRUE.
}

\section{sum_na}{

Calculates the sum of a numeric vector input, with na.rm = TRUE.
}

\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
