% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_format.R
\name{p_format}
\alias{p_format}
\alias{vp_format}
\title{Formats p values for use on plots}
\usage{
p_format(p, pmin = 1e-04, scientific = FALSE)

vp_format(p, pmin = 1e-04, scientific = FALSE)
}
\arguments{
\item{p}{numeric p-value to be formatted}

\item{pmin}{= 0.0001; numeric minimum p value to display.
all values below 0.0001 displayed as '< 0.0001'}

\item{scientific}{= FALSE; logical, whether or not to use scientific notation}
}
\value{
A character vector with the text of the formatted p-values.
The vector will be as long as the number of arguments passed to it.
p_format should be used only for single p-values; vectors of multiple
p-values should use vp_format.
}
\description{
Custom format for p values. Works well with ggpubr's stat_compare_means.
}
\details{
P values are formatted as follows:
  - values below pmin (default is 0.0001) are formatted as 'p < pmin'
  - values between pmin and 0.01 are formatted to have only 1 sigfig
  (i.e. 'p = 0.002')
  - values above 0.01 are formatted to have 2 sigfigs (i.e. 'p = 0.016',
  'p = 0.35')
}
\section{p_format}{

Formats a numeric value as a p value that displays well on plots. When used
on a numeric vector, all values will be formatted in the same manner as the
first value, regardless of what the actual value is. This means that if the
first value is formatted as 'p < 0.0001', all values will be, even if they
are greater than 0.0001. For this reason, numeric vectors should be formatted
using vp_format.
}

\section{vp_format}{

Vectorized version of p_format, that applies the p_format function to each
member of a vector individually. Particularly useful with for faceted plots
in ggpubr::stat_compare_means (see example).
}

\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
\keyword{figure,}
\keyword{format,}
\keyword{format.pval,}
\keyword{ggplot,}
\keyword{ggpubr,}
\keyword{p.format}
\keyword{plot,}
\keyword{pval,}
\keyword{significance,}
