% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_.R
\name{paste_}
\alias{paste_}
\title{Wrapper to paste things together with "_" as the separator.}
\usage{
paste_(...)
}
\arguments{
\item{...}{Arguments to pass to base paste function.}
}
\value{
A string with all elements pasted together with '_' in between them.
}
\description{
Paste things together with '_' as the separator.
}
\details{
Useful for saving file names or automatically generating variable names, etc
}
\author{
Alec Plotkin, \email{alp.plotkin@gmail.com}
}
